refactor: 날짜 포매팅 util로 분리
